refactor: deduplicate shared helpers across @granit packages
Removes structural duplication surfaced by the framework audit (jscpd),
with no behaviour change — public API additions are additive only.

- arch-tests-kit: hoist the byte-identical findSubdirs/processSubdirEntry
  into fs.ts (next to walkSourceFiles); naming.ts and patterns.ts now import
  the single copy and drop their unused node:fs/node:path imports.
- notifications: extract createTransportListeners() into the core package;
  the SignalR and SSE transport adapters delegate the listener-set / setState
  / emit bookkeeping to it instead of each reimplementing the registry.
- react-taxonomy: share DEFAULT_TAG_CHIP_STRIP_LABELS between TagChipStrip
  and DocumentTagChipStrip.

Net -72 lines of duplication. Verified: pnpm tsc + lint + tests (1034) green.
